Patch for scan type definition
Patch: scan_type md is set to scan_nd for all energy scans (because it's set at that level), so I added a new md entry enscan_type that matches the plan name we call.

Question: if we set the scan_type inour scan, will that override the original scan_nd?  If so that's much more elegant.