Sample locate and spiral search enhancements
Fixed typos and added next and previous commands (n/p) to bar imaging utility.

Moved misc bar manipulation commands (straighten, flip, etc) to sample locate since conceptually that seems a better place for them to me.

Added an assert in correct_bar that the lengths on camera vs in chamber have to be within 5 mm of each other, which ought to reduce confusion about where AFs are.

Modified spiralsearch_all: if you ctrl-c a scan, if you do RE.stop(), it will just continue to the next sample.  RE.abort() will stop the whole series.  Hopefully this works- I tested it on my machine, but in a Jupyter nb not an ipython shell.  Some testing at the beamline would be advisable.